+++ lib/ex_icon/template.ex
@@ -41,8 +41,9 @@ defmodule ExIcon.Template do
     error -> {:syntax_error, Exception.message(error)}
   end
 
+  @spec refuse_syntax_error!(binary) :: no_return()
   defp refuse_syntax_error!(message) do
-    raise """
+    Mix.raise("""
     Refusing to write the generated module
 
     The generated code is not valid Elixir:
@@ -50,9 +51,10 @@ defmodule ExIcon.Template do
         #{message}
 
     An icon whose name cannot be used as a function name can cause this.
-    """
+    """)
   end
 
+  @spec refuse_unexpected!() :: no_return()
   defp refuse_unexpected! do
     raise """
     Refusing to write the generated module